% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-beta.R
\name{beta}
\alias{beta}
\title{Regression Coefficients
as a Function of the Covariance Matrix}
\usage{
beta(sigmacapx, sigmayx, verbose = TRUE)
}
\arguments{
\item{sigmacapx}{Numeric matrix.
Covariance matrix of the regressors.}

\item{sigmayx}{Numeric vector.
Covariances between the regressand and regressor variables
\eqn{
    \boldsymbol{\sigma}_{y , \mathbf{X}}
    =
    \{ \sigma_{y, x_1}, \sigma_{y, x_j}, \sigma_{y, x_p} \}^{\prime}
  }
where
\eqn{j = \{ 1, \cdots, p \}}.}

\item{verbose}{Logical.
If \code{verbose = TRUE}, print message if error occurs.}
}
\value{
A numeric vector.
}
\description{
Regression Coefficients
as a Function of the Covariance Matrix
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{beta_of_vechsigmacap}()},
\code{\link{beta_wrt_vechsigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{moments_helper}()},
\code{\link{mu_of_theta}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq_of_theta}()},
\code{\link{rsq}()},
\code{\link{sigmacap_of_thetastar}()},
\code{\link{sigmacap_of_theta}()},
\code{\link{sigmasq_of_vechsigmacap}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()},
\code{\link{theta_helper}()},
\code{\link{theta_of_moments}()},
\code{\link{theta_of_sigmacap}()},
\code{\link{theta_of_vechsigmacap}()},
\code{\link{theta_wrt_vechsigmacap}()},
\code{\link{thetastar_helper}()},
\code{\link{thetastar_of_sigmacap}()},
\code{\link{thetastar_of_vechsigmacap}()},
\code{\link{thetastar}()},
\code{\link{theta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
