% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-betastar_of_theta.R
\name{betastar_of_theta}
\alias{betastar_of_theta}
\title{Standardized Regression Coefficients
as a Function of the Vector of Regression Parameters}
\usage{
betastar_of_theta(x)
}
\arguments{
\item{x}{Numeric vector.
Vector of parameters
\eqn{\boldsymbol{theta}
  =
  \{
    \boldsymbol{\beta},
    \sigma^2,
    \mathrm{\vech}
    \left( \boldsymbol{\Sigma}_{\mathbf{X}, \mathbf{X}} \right)
  \}^{\prime}
  }.}
}
\value{
A numeric vector.
}
\description{
Standardized Regression Coefficients
as a Function of the Vector of Regression Parameters
}
\section{Dependencies}{
\itemize{
\item \code{\link[=sym_of_vech]{sym_of_vech()}}
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
