% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-betastar_of_vechsigmacap.R
\name{betastar_of_vechsigmacap}
\alias{betastar_of_vechsigmacap}
\title{Standardized Regression Coefficients
as a Function of the Half-Vectorization
of the Covariance Matrix}
\usage{
betastar_of_vechsigmacap(x)
}
\arguments{
\item{x}{Numeric vector.
Half-vectorization of the covariance matrix of
\eqn{\{y, x_1, \cdots, x_p \}^{\prime}}.}
}
\value{
A numeric vector.
}
\description{
Standardized Regression Coefficients
as a Function of the Half-Vectorization
of the Covariance Matrix
}
\section{Dependencies}{
\itemize{
\item \code{\link[=sym_of_vech]{sym_of_vech()}}
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
