% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-dcap.R
\name{dcap}
\alias{dcap}
\title{Duplication Matrix}
\usage{
dcap(x)
}
\arguments{
\item{x}{Integer.
Dimension of the symmetric matrix.}
}
\value{
A matrix.
}
\description{
Create a duplication matrix.
}
\details{
The duplication matrix \eqn{D_{m}} is the
\eqn{
  m^2 \times \frac{m \left( m + 1 \right)}{2}
}
matrix
for a given \eqn{m \times m} symmetric matrix \eqn{A}
where
\eqn{
  D_{m} \mathrm{vech} \left( A \right)
  =
  \mathrm{vec} \left( A \right)
}.
}
\examples{
dcap(3)
}
\references{
\href{https://en.wikipedia.org/wiki/Duplication_and_elimination_matrices#Duplication_matrix}{Wikipedia: Duplication matrix}
}
\seealso{
Other Linear Algebra Functions: 
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Linear Algebra Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
