% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-moments_helper.R
\name{moments_helper}
\alias{moments_helper}
\title{Helper Function to Convert a Vector of Moments
to the Mean Vector and Covariance Matrix}
\usage{
moments_helper(x)
}
\arguments{
\item{x}{Numeric vector.
Vector of parameters
\eqn{
    \boldsymbol{\theta}
    =
    \left\{
      \boldsymbol{\mu} ,
      \mathrm{vech} \left( \boldsymbol{\Sigma} \right)
    \right\}^{\prime}
  }.}
}
\value{
A list with elements \code{mu} and \code{sigmacap}.
}
\description{
Helper Function to Convert a Vector of Moments
to the Mean Vector and Covariance Matrix
}
\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{beta_of_vechsigmacap}()},
\code{\link{beta_wrt_vechsigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{beta}()},
\code{\link{mu_of_theta}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq_of_theta}()},
\code{\link{rsq}()},
\code{\link{sigmacap_of_thetastar}()},
\code{\link{sigmacap_of_theta}()},
\code{\link{sigmasq_of_vechsigmacap}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()},
\code{\link{theta_helper}()},
\code{\link{theta_of_moments}()},
\code{\link{theta_of_sigmacap}()},
\code{\link{theta_of_vechsigmacap}()},
\code{\link{theta_wrt_vechsigmacap}()},
\code{\link{thetastar_helper}()},
\code{\link{thetastar_of_sigmacap}()},
\code{\link{thetastar_of_vechsigmacap}()},
\code{\link{thetastar}()},
\code{\link{theta}()}
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
