% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-mu_of_theta.R
\name{mu_of_theta}
\alias{mu_of_theta}
\title{Mean vector as a Function of Regression Parameters}
\usage{
mu_of_theta(x)
}
\arguments{
\item{x}{Numeric vector.
Vector of parameters
\eqn{
    \boldsymbol{\theta}
    =
    \left\{
      \boldsymbol{\mu} ,
      \mathrm{vech} \left( \boldsymbol{\Sigma} \right)
    \right\}^{\prime}
  }.}
}
\value{
A numeric vector.
}
\description{
Mean vector as a Function of Regression Parameters
}
\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{beta_of_vechsigmacap}()},
\code{\link{beta_wrt_vechsigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{beta}()},
\code{\link{moments_helper}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq_of_theta}()},
\code{\link{rsq}()},
\code{\link{sigmacap_of_thetastar}()},
\code{\link{sigmacap_of_theta}()},
\code{\link{sigmasq_of_vechsigmacap}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()},
\code{\link{theta_helper}()},
\code{\link{theta_of_moments}()},
\code{\link{theta_of_sigmacap}()},
\code{\link{theta_of_vechsigmacap}()},
\code{\link{theta_wrt_vechsigmacap}()},
\code{\link{thetastar_helper}()},
\code{\link{thetastar_of_sigmacap}()},
\code{\link{thetastar}()},
\code{\link{theta}()}
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
