% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rmvn_chol.R
\name{rmvn_chol}
\alias{rmvn_chol}
\title{Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition}
\usage{
rmvn_chol(n, mu, sigmacap, varnames = NULL, data_frame = FALSE)
}
\arguments{
\item{n}{Positive integer.
\code{n} variates.}

\item{mu}{Numeric vector.
Parameter.
Mean vector
\eqn{\boldsymbol{\mu}}.}

\item{sigmacap}{Numeric matrix.
Parameter.
Covariance matrix
\eqn{\boldsymbol{\Sigma}}.}

\item{varnames}{Character vector
Optional variable names.}

\item{data_frame}{Logical.
If \code{data_frame = TRUE},
returns a \code{data.frame}.
If \code{data_frame = FALSE},
returns a \code{matrix}.}
}
\value{
A matrix (\code{data_frame = FALSE}) or data.frame (\code{data_frame = TRUE}).
}
\description{
Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition
}
\examples{
mu <- c(100, 100)
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

x <- rmvn_chol(
  n = 100,
  mu = mu,
  sigmacap = sigmacap
)

colMeans(x)
cov(x)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
