% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-rsq.R
\name{rsq}
\alias{rsq}
\title{R-Squared as a Function of Regression Parameters}
\usage{
rsq(beta, sigmasq, sigmayx, sigmacapx)
}
\arguments{
\item{beta}{Numeric vector.
Partial regression slopes
\eqn{
    \boldsymbol{\beta}
    =
    \{ \beta_1, \cdots, \beta_p \}^{\prime}
  }.}

\item{sigmasq}{Numeric.
Error variance
\eqn{\sigma^2}.}

\item{sigmayx}{Numeric vector.
Covariances between the regressand and regressor variables
\eqn{
    \boldsymbol{\sigma}_{y , \mathbf{X}}
    =
    \{ \sigma_{y, x_1}, \sigma_{y, x_j}, \sigma_{y, x_p} \}^{\prime}
  }
where
\eqn{j = \{ 1, \cdots, p \}}.}

\item{sigmacapx}{Numeric matrix.
Covariance matrix of the regressor variables
\eqn{
    \boldsymbol{\Sigma}_{\mathrm{X}, \mathrm{X}}
  }.}
}
\value{
A numeric vector of length 1.
}
\description{
R-Squared as a Function of Regression Parameters
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{beta_of_vechsigmacap}()},
\code{\link{beta_wrt_vechsigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{beta}()},
\code{\link{moments_helper}()},
\code{\link{mu_of_theta}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq_of_theta}()},
\code{\link{sigmacap_of_thetastar}()},
\code{\link{sigmacap_of_theta}()},
\code{\link{sigmasq_of_vechsigmacap}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()},
\code{\link{theta_helper}()},
\code{\link{theta_of_moments}()},
\code{\link{theta_of_sigmacap}()},
\code{\link{theta_of_vechsigmacap}()},
\code{\link{theta_wrt_vechsigmacap}()},
\code{\link{thetastar_helper}()},
\code{\link{thetastar_of_sigmacap}()},
\code{\link{thetastar_of_vechsigmacap}()},
\code{\link{thetastar}()},
\code{\link{theta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
