% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-rsq_of_sigmacap.R
\name{rsq_of_sigmacap}
\alias{rsq_of_sigmacap}
\title{R-Squared as a Function of the Covariance Matrix}
\usage{
rsq_of_sigmacap(x)
}
\arguments{
\item{x}{Numeric matrix.
Covariance matrix
\eqn{\boldsymbol{\Sigma}}
of
\eqn{\{y, x_1, \cdots, x_p \}^{\prime}}.}
}
\value{
A numeric vector of length 1.
}
\description{
R-Squared as a Function of the Covariance Matrix
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{rsq}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
