% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-rsq_of_theta.R
\name{rsq_of_theta}
\alias{rsq_of_theta}
\title{R-Squared as a Function
the Vector of Regression Parameters}
\usage{
rsq_of_theta(x)
}
\arguments{
\item{x}{Numeric vector.
Vector of parameters
\eqn{\boldsymbol{theta}
  =
  \{
    \boldsymbol{\beta},
    \sigma^2,
    \mathrm{\vech}
    \left( \boldsymbol{\Sigma}_{\mathbf{X}, \mathbf{X}} \right)
  \}^{\prime}
  } when \code{mean_structure = FALSE} or
\eqn{\boldsymbol{theta}
  =
  \{
    \boldsymbol{\beta},
    \sigma^2,
    \mathrm{\vech}
    \left( \boldsymbol{\Sigma}_{\mathbf{X}, \mathbf{X}} \right),
    \alpha,
    \boldsymbol{\mu}_{\mathbf{X}}
  \}^{\prime}
  } when \code{mean_structure = TRUE}}
}
\value{
A numeric vector of length 1.
}
\description{
R-Squared as a Function
the Vector of Regression Parameters
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{beta_of_vechsigmacap}()},
\code{\link{beta_wrt_vechsigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{beta}()},
\code{\link{moments_helper}()},
\code{\link{mu_of_theta}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq}()},
\code{\link{sigmacap_of_thetastar}()},
\code{\link{sigmacap_of_theta}()},
\code{\link{sigmasq_of_vechsigmacap}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()},
\code{\link{theta_helper}()},
\code{\link{theta_of_moments}()},
\code{\link{theta_of_sigmacap}()},
\code{\link{theta_of_vechsigmacap}()},
\code{\link{theta_wrt_vechsigmacap}()},
\code{\link{thetastar_helper}()},
\code{\link{thetastar_of_sigmacap}()},
\code{\link{thetastar}()},
\code{\link{theta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
