% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-sigmaysq.R
\name{sigmaysq}
\alias{sigmaysq}
\title{Variance of the Regressand Variable}
\usage{
sigmaysq(sigmasq, beta, sigmacapx)
}
\arguments{
\item{sigmasq}{Numeric.
Error variance
\eqn{\sigma^2}.}

\item{beta}{Numeric vector.
Partial regression slopes
\eqn{
    \boldsymbol{\beta}
    =
    \{ \beta_1, \cdots, \beta_p \}^{\prime}
  }.}

\item{sigmacapx}{Numeric matrix.
Covariance matrix of the regressor variables
\eqn{
    \boldsymbol{\Sigma}_{\mathrm{X}, \mathrm{X}}
  }.}
}
\value{
A numeric vector of length 1.
}
\description{
Variance of the Regressand Variable
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq}()},
\code{\link{sigmasq}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
