% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-package.R
\docType{package}
\name{strRegression}
\alias{strRegression}
\title{strRegression: Structure of Regression}
\description{
A collection of functions
related to the mean and covariance structures
of the linear regression model.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{package}
\keyword{strRegression}
