% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-sym_of_vech.R
\name{sym_of_vech}
\alias{sym_of_vech}
\title{Symmetric matrix A from vech(A)}
\usage{
sym_of_vech(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
A vector.
}
\description{
Symmetric matrix from its half-vectorization.
}
\details{
Generates an \eqn{m \times m} symmetric matrix
from an \eqn{m(m + 1) / 2} vector.
}
\examples{
A <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  ncol = 3
)
vechA <- c(1.0, 0.5, 0.4, 1.0, 0.6, 1.0)

sym_of_vech(vechA)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)#Half-vectorization}{Wikipedia: Half-vectorization}
}
\seealso{
Other Linear Algebra Functions: 
\code{\link{vechnames}()},
\code{\link{vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Linear Algebra Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
