% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-theta_wrt_vechsigmacap.R
\name{theta_wrt_vechsigmacap}
\alias{theta_wrt_vechsigmacap}
\title{Jacobian Matrix of the Regression Parameters
with Respect to the Unique Elements of the Covariance Matrix}
\usage{
theta_wrt_vechsigmacap(x)
}
\arguments{
\item{x}{Numeric vector.
Half-vectorization of the covariance matrix
\eqn{\boldsymbol{\Sigma}}
of
\eqn{\left\{y, x_1, \dots, x_p \right\}^{\prime}}.}
}
\value{
A matrix.
}
\description{
Matrix of partial derivatives of unstandardized parameters
with respect to the unique elements of the covariance matrix.
Note that the parameter vector is ordered as
\deqn{
  \boldsymbol{\theta}
  =
  \left\{
  \boldsymbol{\beta},
  \sigma^{2},
  \mathrm{vech} \left( \boldsymbol{\Sigma}_{\mathbf{X}, \mathbf{X}} \right)
  \right\}
}
and the covariance matrix \eqn{\boldsymbol{\Sigma}} is ordered as
\deqn{
  \left\{
  y,
  x_1,
  \dots,
  x_p
  \right\}^{\prime}
}.
}
\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{beta_of_vechsigmacap}()},
\code{\link{beta_wrt_vechsigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{beta}()},
\code{\link{moments_helper}()},
\code{\link{mu_of_theta}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq_of_theta}()},
\code{\link{rsq}()},
\code{\link{sigmacap_of_thetastar}()},
\code{\link{sigmacap_of_theta}()},
\code{\link{sigmasq_of_vechsigmacap}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()},
\code{\link{theta_helper}()},
\code{\link{theta_of_moments}()},
\code{\link{theta_of_sigmacap}()},
\code{\link{theta_of_vechsigmacap}()},
\code{\link{thetastar_helper}()},
\code{\link{thetastar_of_sigmacap}()},
\code{\link{thetastar_of_vechsigmacap}()},
\code{\link{thetastar}()},
\code{\link{theta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
