% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-thetastar_helper.R
\name{thetastar_helper}
\alias{thetastar_helper}
\title{Helper Function to Convert a Vector of Standardized Parameters
to a List of Standardized Parameters}
\usage{
thetastar_helper(x)
}
\arguments{
\item{x}{Numeric vector.
Vector of standardized parameters
\eqn{\boldsymbol{theta}^{\ast}
  =
  \{
    \boldsymbol{\beta}^{\ast},
    \sigma_{y},
    \boldsymbol{\sigma}_{\mathbf{X}},
    \mathrm{\vechs}
    \left( \mathbf{P}_{\mathbf{X}, \mathbf{X}} \right)
  \}^{\prime}
  }}
}
\value{
A list.
}
\description{
Helper Function to Convert a Vector of Standardized Parameters
to a List of Standardized Parameters
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{beta_of_vechsigmacap}()},
\code{\link{beta_wrt_vechsigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{beta}()},
\code{\link{moments_helper}()},
\code{\link{mu_of_theta}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq_of_theta}()},
\code{\link{rsq}()},
\code{\link{sigmacap_of_thetastar}()},
\code{\link{sigmacap_of_theta}()},
\code{\link{sigmasq_of_vechsigmacap}()},
\code{\link{sigmasq}()},
\code{\link{sigmaysq}()},
\code{\link{theta_helper}()},
\code{\link{theta_of_moments}()},
\code{\link{theta_of_sigmacap}()},
\code{\link{theta_of_vechsigmacap}()},
\code{\link{theta_wrt_vechsigmacap}()},
\code{\link{thetastar_of_sigmacap}()},
\code{\link{thetastar_of_vechsigmacap}()},
\code{\link{thetastar}()},
\code{\link{theta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
