% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vech.R
\name{vech}
\alias{vech}
\title{Half-Vectorize}
\usage{
vech(x, names = FALSE, sep = ".")
}
\arguments{
\item{x}{Matrix.}

\item{names}{Logical.
Add names.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A vector.
}
\description{
Half-vectorize a symmetric matrix.
}
\details{
Generates an \eqn{m(m + 1) / 2} vector
from the unique elements
of an \eqn{m \times m} symmetric matrix
by stacking the columns (column-major).
}
\section{Dependencies}{
\itemize{
\item \code{\link[=vechnames]{vechnames()}}
}
}

\examples{
x <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  ncol = 3
)

vech(x)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)#Half-vectorization}{Wikipedia: Half-vectorization}
}
\seealso{
Other Linear Algebra Functions: 
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Linear Algebra Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
\keyword{vectorization}
