% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-vechsnames.R
\name{vechsnames}
\alias{vechsnames}
\title{Vector Names for Strict Half-Vectorization}
\usage{
vechsnames(x, sep = ".")
}
\arguments{
\item{x}{Character vector of names.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A vector.
}
\description{
Vector Names for Strict Half-Vectorization
}
\examples{
x <- diag(2)
colnames(x) <- rownames(x) <- c("x1", "x2")
vechsnames(colnames(x))

x <- diag(3)
colnames(x) <- rownames(x) <- c("x1", "x2", "x3")
vechsnames(colnames(x))
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)#Half-vectorization}{Wikipedia: Half-vectorization}
}
\seealso{
Other Linear Algebra Functions: 
\code{\link{dcap}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()},
\code{\link{vechnames}()},
\code{\link{vechs}()},
\code{\link{vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Linear Algebra Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
\keyword{vectorization}
