% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRegression-sigmasq.R
\name{sigmasq}
\alias{sigmasq}
\title{Error Variance}
\usage{
sigmasq(sigmaysq, beta, sigmacapx)
}
\arguments{
\item{sigmaysq}{Numeric.
Variance of the regressand variable
\eqn{\sigma_{y}^{2}}.}

\item{beta}{Numeric vector.
Partial regression slopes
\eqn{
    \boldsymbol{\beta}
    =
    \{ \beta_1, \cdots, \beta_p \}^{\prime}
  }.}

\item{sigmacapx}{Numeric matrix.
Covariance matrix of the regressor variables
\eqn{
    \boldsymbol{\Sigma}_{\mathrm{X}, \mathrm{X}}
  }.}
}
\value{
A numeric vector of length 1.
}
\description{
Error Variance
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\seealso{
Other Structure of Regression Functions: 
\code{\link{beta_of_sigmacap}()},
\code{\link{betastar_of_rhocap}()},
\code{\link{betastar_of_sigmacap}()},
\code{\link{betastar_of_theta}()},
\code{\link{betastar_of_vechsigmacap}()},
\code{\link{moments_helper}()},
\code{\link{rsq_of_sigmacap}()},
\code{\link{rsq}()},
\code{\link{sigmaysq}()},
\code{\link{theta_helper}()},
\code{\link{theta_of_moments}()},
\code{\link{theta_of_sigmacap}()},
\code{\link{theta_of_vechsigmacap}()},
\code{\link{thetastar_of_sigmacap}()},
\code{\link{thetastar}()},
\code{\link{theta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Structure of Regression Functions}
\keyword{strRegression}
